requirements.txt
# Core framework
Django==4.2.0

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1

# Database driver
mysqlclient==2.2.0

# External service integrations
requests==2.31.0

# Image handling
Pillow==10.1.0
